refactor(frontend): rename spiderfy.ts → fan-layout.ts + gate window.__birdMap (#290, #291)
leaving spiderfy.ts as a misnomer (pure geometry + paint constants).
Rename to fan-layout.ts. Drop computePrePanOffset (only spiderfy.test
imported it; pre-pan was a click-spiderfy concern). Keep
buildSpiderfyLeaderLineFeatures — auto-spider still uses it.

handleLoad — runs in production too, exposing the maplibre instance
to anyone running the property in devtools. Gate to non-prod via
import.meta.env.MODE. Add unmount cleanup so a remount doesn't leave
a stale handle behind.
